+++ src/pools/abstract-pool.ts
@@ -117,6 +117,10 @@ export abstract class AbstractPool<
    */
   private destroying: boolean
   /**
+   * Whether the pool ready event has been emitted or not.
+   */
+  private readyEventEmitted: boolean
+  /**
    * The start timestamp of the pool.
    */
   private readonly startTimestamp
@@ -166,6 +170,7 @@ export abstract class AbstractPool<
     this.started = false
     this.starting = false
     this.destroying = false
+    this.readyEventEmitted = false
     if (this.opts.startWorkers === true) {
       this.start()
     }
@@ -982,6 +987,7 @@ export abstract class AbstractPool<
       }),
     )
     this.emitter?.emit(PoolEvents.destroy, this.info)
+    this.readyEventEmitted = false
     this.destroying = false
     this.started = false
   }
@@ -1573,12 +1579,9 @@ export abstract class AbstractPool<
     )
     workerInfo.ready = message.ready as boolean
     workerInfo.taskFunctionNames = message.taskFunctionNames
-    if (this.ready) {
-      const emitPoolReadyEventOnce = once(
-        () => this.emitter?.emit(PoolEvents.ready, this.info),
-        this,
-      )
-      emitPoolReadyEventOnce()
+    if (!this.readyEventEmitted && this.ready) {
+      this.readyEventEmitted = true
+      this.emitter?.emit(PoolEvents.ready, this.info)
     }
   }