+++ tests/pools/abstract-pool.test.mjs
@@ -1082,6 +1082,7 @@ Deno.test({
         )
         expect(pool.info.started).toBe(false)
         expect(pool.info.ready).toBe(false)
+        expect(pool.readyEventEmitted).toBe(false)
         expect(pool.workerNodes).toStrictEqual([])
         await expect(pool.execute()).rejects.toThrow(
           new Error('Cannot execute a task on not started pool'),
@@ -1089,6 +1090,8 @@ Deno.test({
         pool.start()
         expect(pool.info.started).toBe(true)
         expect(pool.info.ready).toBe(true)
+        await waitPoolEvents(pool, PoolEvents.ready, 1)
+        expect(pool.readyEventEmitted).toBe(true)
         expect(pool.workerNodes.length).toBe(numberOfWorkers)
         for (const workerNode of pool.workerNodes) {
           expect(workerNode).toBeInstanceOf(WorkerNode)